+++ core/src/main/java/io/cucumber/core/api/event/HookType.java
@@ -1,10 +1,12 @@
 package io.cucumber.core.api.event;
 
+import static java.util.Locale.ROOT;
+
 public enum HookType {
     Before, After, BeforeStep, AfterStep;
 
     @Override
     public String toString() {
-        return super.toString().toLowerCase();
+        return super.toString().toLowerCase(ROOT);
     }
 }