+++ core/src/main/java/io/cucumber/core/stepexpression/TypeResolver.java
@@ -8,9 +8,9 @@ import java.lang.reflect.Type;
 public interface TypeResolver {
 
     /**
-     * A type to data convert the table or doc string to.
-     *
-     * May return null if no type could be resolved.
+     * A type to data convert the table or doc string to. May not return null.
+     * <p>
+     * When the {@link Object} type is returned no transform will be applied.
      *
      * @return a type
      */