+++ java/src/main/java/io/cucumber/java/JavaStepDefinition.java
@@ -17,6 +17,7 @@ import java.lang.annotation.Annotation;
 import java.lang.reflect.Method;
 import java.lang.reflect.Type;
 import java.util.ArrayList;
+import java.lang.reflect.Type;
 import java.util.List;
 
 class JavaStepDefinition implements StepDefinition {
@@ -54,7 +55,11 @@ class JavaStepDefinition implements StepDefinition {
 
     public List<Argument> matchedArguments(PickleStep step) {
         ArgumentMatcher argumentMatcher = new ExpressionArgumentMatcher(expression);
-        return argumentMatcher.argumentsFrom(step);
+        Type[] types = new Type[parameterInfos.size()];
+        for (int i = 0; i < types.length; i++) {
+            types[i] = parameterInfos.get(i).getType();
+        }
+        return argumentMatcher.argumentsFrom(step, types);
     }
 
     public String getLocation(boolean detail) {